requirements.txt
asgiref==3.3.1
attrs==20.3.0
Django==3.1.4
django-ckeditor==6.0.0
django-environ==0.4.5
django-js-asset==1.2.2
djangorestframework==3.12.2
Faker==5.0.2
iniconfig==1.1.1
packaging==20.8
pluggy==0.13.1
py==1.10.0
PyJWT==2.0.0
pyparsing==2.4.7
pytest==6.2.1
pytest-django==4.1.0
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
